Update patch set 5
Patch Set 5:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Already tried that. I also tried to simplify a lot this code. Nothing worked
> 
> but this PS3 fixes it?

Using PS3-5 it even fixes it on the green color, but that white one seems to be more reliable. I dont like it either but for now we have to use it sadly as cleaning it up or changing its order breaks it again.

Patch-set: 5
Reviewer: Gerrit User 1000143 <1000143@3162e3d8-d894-470e-abd9-724f7b0effad>