Update patch set 1
Patch Set 1: Code-Review-2

Whitelist is there for a reason.
We tried forcing AAC a long time ago and it broke a whole lot of SBC-only devices.

MIUI allows users to set device to negotiate with AAC manually.

Even if this patch turns out to be fine, it should be on the global property, not per-device one.

Patch-set: 1
Reviewer: Gerrit User 1000062 <1000062@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Code-Review=-2