Update patch set 1
Patch Set 1:

The first push was here: https://gerrit.aospa.co/#/c/12974 and was later abandoned as it was merged in a different commit, squashed

I took it, gave it a little nicer and more accurate description, and pushed it as it worked for our testers. None of the testers reported issues, but that said, it's a small group.

Patch-set: 1
Reviewer: Gerrit User 1000541 <1000541@3162e3d8-d894-470e-abd9-724f7b0effad>